Change "Delete" to "Remove" for cell editing
The confirmation button that appears when a cell is deleted now says
"Remove" instead of "Delete" (or the localized equivalent of "Delete").

This completes the chapter 9 bronze challenge.  However, this is done
using new UIKit methods and classes that were not added until iOS 8.  I
wonder this is done prior to iOS 8.