Allow user-edited item photos
This is more like "force user-edited item photos".
The user is presented with a crop box, that can't be adjusted and is
mostly just annoying, after the user takes a photo.  The edited photo
is what is stored in the ImageStore for the item.

Why do this?  I guess it makes for nice, square images. It also
completes the Chapter 11 Bronze challenge.