Add crosshair overlay view to ImagePicker
The cross is not quite in the center of the image preview.  Not sure
how to get a reference -- in a clean way that isn't gleaned from debug
information and is more app store friendly -- to the preview pane. So
the cross hair is in the center of the screen, which is slightly off
center with respect to the image preview.